revised the multiple choice code and add a tests
also added go modules

the test currently fails as it would when running student code.
this is not ideal for testing purposes, but the code for the
multiple choice is a bit hard to test in its current form.
a revision shouldn't work on the individual file. it would
probably be easier to work with a slice of strings (lines).
but don't have time now to fix this.